+++ _meta.lua
@@ -3,5 +3,5 @@ return {
   name = "livelib",
   fullname = _("Livelib"),
   description = _([[Sync reading status and ratings with livelib.ru]]),
-  version = "0.4.0",
+  version = "0.5.0",
 }